Test Markdown documentation
Adds some pytest hook code to "run" our Markdown example
documentation.

This will take each Markdown file in the `docs/` directory, parse it,
and run all code blocks in each file in order. The stdout of each code
block is then compared to an expected value in an optional following
`{testoutput}` block. If the output is missing, the output is assumed
to be nothing.