Adds `mp_ctx` arg to entry points which spawn subprocesses
Provide a way for the calling code to specify which "multiprocessing
context" to use to spawn subprocesses. See
https://docs.python.org/3/library/multiprocessing.html#contexts-and-start-methods

I'm using this to allow us to mock out multiprocessing with
multithreading in doctests. This will also let you more easily test
differences between "spawn" and "fork" modes.

I'm defaulting to using "spawn" because I think "fork" mode was the
cause of some mysterious hanging in tests. General consensus seems to
be "spawn" is less buggy:
https://github.com/python/cpython/issues/84559 I've felt like tests
are consistently faster with it.

Also uses the `multiprocessing.Manager` as a context manager so it
gets cleaned up correctly. This might have been the cause of other
hanging in local cluster execution.