Windowing
I am incapable of making a PR with less than 1500 lines of code.

Docstrings explain what is going on. Start with `bytewax.window` for
an overview of windowing concepts. Then read the Rust code in the
`window` module for how that's implemented.

There are a few changes to the recovery system you can see in diffs to
the Rust `recovery` module. The main thing is now we mostly are
passing around backups of `Vec<u8>` rather than explicit types so we
can mush all the backup types together.

The most horrible part of all the code is the renaming / re-invention
of the `StatefulMap` operator into `StatefulUnary` which very very
carefully lets you run arbitrary logic on incoming values, keeps track
of state per key, lets you "awaken" the operator per-key even if
there's no data, and in all that manages to send out state snapshots.

I disabled markdown tests until we figure out what we want for input
and how much we want to expose epochs, then we can turn them back
on. There's one pytest for windowing now, though.

The `KafkaPump` has been slightly modified to emit epochs periodically
and not deal with batch sizes, since we're moving in the direction of
heartbeat epochs.