bugfix: Kafka AdminClient problems
Locally, when connecting to kafka, I was getting an error that
there were no partition counts in the Kafka topic description.

I was able to replicate this issue in Python.

After looking through the rdkafka examples, I found a test
that uses a consumer to fetch metadata about a topic.

I've adapted that code to work here instead of using the AdminClient.