Replace generic epoch with explicit u64
We had a ton of generic `<T>` which I originally thought we might want
to use in case we wanted to swithc the type of our epochs. But it
turns out for recovery we have to do some epoch arithmetic so we were
effetively pinning it to `u64` anyway. This gets rid of that
complexity and hard codes that type.