Partitioned output
Refactors output system to look like the partitioned input system.

Python-side, there's a `PartOutput` and `DynamicOutput` that are the
two base output types. All output types derive from them. There are
now two output operators, `part_output` and `dynamic_output` which you
use to add an output at that point in the dataflow. They are two
separate operators because they have different semantics and input
data shapes.

Partitioned output is stateful and has the cluster list all partitions
then will divvy them up to workers. Since consistent routing of data
is crucial, the input to the `part_output` operator needs to look like
`(key, value)` 2-tuples and the operator will route to a consistent
partition. Since each partition is independent, we can use the
standard pattern of having recovery state that is routed back to the
constructor of that partition on resume.

Dynamic output is stateless output that runs on each worker in the
current cluster. We explicitly say "we can't store state" and so make
it obvious that you can not reach the same level of delivery
guarantees as with the other output forms. Since the semantics of how
to key data happen somewhere downstream, the input to the
`dynamic_output` operator has no constraints, items are written
directly to the output on the current worker.

Re-writes all of the existing output config types to use one of these
two base types. Perhaps counter-intuitively `KafkaOutput` is a
`DynamicOutput` because the hashing to partitions happens via the
producer and there are no limit on how many producers you can have.

Rust-side, tries getting rid of layers of abstraction and ends up
taking the Python output definitions directly into Timely
operators. This feels right because we directly want different
dataflow semantics (aka Timely operators) for each type of output.