More deterministic windowing
As part of debugging what was going on with windowing in the
possibly-buggy recovery example, I found some issues with how clocks
and windowers were implemented. This was basically a week of
continually finding more and more obscure off-by-one errors.

First `start_at` is renamed to `align_to`. That former name
incorrectly implied that there were no windows before that time, but
that is not true; windows extend (almost) infinitely in both
directions in time from that alignment point. What determines if an
item is late for a window is not the windowing definition, but the
watermark of the clock.

Next, both `{Sliding,Tumbling}Window` are required to take
`align_to`. Previously, this was filled with the timestamp of the
start of the dataflow, but we did not persist it as the recovery
state, nor can we actually properly persist it because recovery state
is per-state-key but at that point we don't have any key. I didn't
want to work on figuring out a way to store "dataflow global" recovery
state, so to make it easier we require you to pick a fixed alignment
time. If you don't ensure `align_to` is consistent across executions,
you have window boundaries jumping around even when you are using
event time.

Then fixes up and adds additional tests for event time watermarking. I
don't belive we were properly advancing the watermark, preventing it
from going backwards, and handling the extreme no data and EOF cases.

Also does a fix for system time watermarking. Once the data has
completed, we should forever return the watermark as "end of time",
not just when there is an item being processed. I don't believe this
causes any difference in behavior with the way our operators are
currently written, but if we didn't process all available data on each
awakening, this is needed.

Then writes a ton more tests for sliding windowing. I believe we were
improperly mixing in the watermark time into the window
definitions. The definition of the window boundaries is a function of
the arguments to the config, not the current item time or
watermark. The watermark is just used to determine if an item is late.